LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-003-expected.txt


PASS lspace/voffset shifts don't affect mpadded preferred width
PASS lspace/voffset shifts don't affect mpadded size
PASS content is shifted by the specified lspace/voffset
PASS content is shifted by the specified lspace/voffset (RTL)
PASS dynamic changes of lspace/voffset
